+++ src/main/kotlin/net/multigesture/kanama/api/PhysicsDirectSpaceState2D.kt
@@ -22,8 +22,8 @@ open class PhysicsDirectSpaceState2D(handle: MemorySegment) : GodotObject(handle
      *
      * Generated from Godot docs: PhysicsDirectSpaceState2D.intersect_point
      */
-    fun intersectPoint(parameters: PhysicsPointQueryParameters2D?, maxResults: Int = 32): List<Map<String, Any?>> {
-        return ObjectCalls.ptrcallWithObjectAndIntArgRetDictionaryList(intersectPointBind, handle, parameters?.requireOpenHandle() ?: MemorySegment.NULL, maxResults)
+    fun intersectPoint(parameters: PhysicsPointQueryParameters2D, maxResults: Int = 32): List<Map<String, Any?>> {
+        return ObjectCalls.ptrcallWithObjectAndIntArgRetDictionaryList(intersectPointBind, handle, parameters.requireOpenHandle(), maxResults)
     }
 
     /**
@@ -38,8 +38,8 @@ open class PhysicsDirectSpaceState2D(handle: MemorySegment) : GodotObject(handle
      *
      * Generated from Godot docs: PhysicsDirectSpaceState2D.intersect_ray
      */
-    fun intersectRay(parameters: PhysicsRayQueryParameters2D?): Map<String, Any?> {
-        return ObjectCalls.ptrcallWithObjectArgRetDictionary(intersectRayBind, handle, parameters?.requireOpenHandle() ?: MemorySegment.NULL)
+    fun intersectRay(parameters: PhysicsRayQueryParameters2D): Map<String, Any?> {
+        return ObjectCalls.ptrcallWithObjectArgRetDictionary(intersectRayBind, handle, parameters.requireOpenHandle())
     }
 
     /**
@@ -52,8 +52,8 @@ open class PhysicsDirectSpaceState2D(handle: MemorySegment) : GodotObject(handle
      *
      * Generated from Godot docs: PhysicsDirectSpaceState2D.intersect_shape
      */
-    fun intersectShape(parameters: PhysicsShapeQueryParameters2D?, maxResults: Int = 32): List<Map<String, Any?>> {
-        return ObjectCalls.ptrcallWithObjectAndIntArgRetDictionaryList(intersectShapeBind, handle, parameters?.requireOpenHandle() ?: MemorySegment.NULL, maxResults)
+    fun intersectShape(parameters: PhysicsShapeQueryParameters2D, maxResults: Int = 32): List<Map<String, Any?>> {
+        return ObjectCalls.ptrcallWithObjectAndIntArgRetDictionaryList(intersectShapeBind, handle, parameters.requireOpenHandle(), maxResults)
     }
 
     /**
@@ -68,8 +68,8 @@ open class PhysicsDirectSpaceState2D(handle: MemorySegment) : GodotObject(handle
      *
      * Generated from Godot docs: PhysicsDirectSpaceState2D.cast_motion
      */
-    fun castMotion(parameters: PhysicsShapeQueryParameters2D?): List<Float> {
-        return ObjectCalls.ptrcallWithObjectArgRetPackedFloat32List(castMotionBind, handle, parameters?.requireOpenHandle() ?: MemorySegment.NULL)
+    fun castMotion(parameters: PhysicsShapeQueryParameters2D): List<Float> {
+        return ObjectCalls.ptrcallWithObjectArgRetPackedFloat32List(castMotionBind, handle, parameters.requireOpenHandle())
     }
 
     /**
@@ -82,8 +82,8 @@ open class PhysicsDirectSpaceState2D(handle: MemorySegment) : GodotObject(handle
      *
      * Generated from Godot docs: PhysicsDirectSpaceState2D.collide_shape
      */
-    fun collideShape(parameters: PhysicsShapeQueryParameters2D?, maxResults: Int = 32): List<Vector2> {
-        return ObjectCalls.ptrcallWithObjectAndIntArgRetVector2List(collideShapeBind, handle, parameters?.requireOpenHandle() ?: MemorySegment.NULL, maxResults)
+    fun collideShape(parameters: PhysicsShapeQueryParameters2D, maxResults: Int = 32): List<Vector2> {
+        return ObjectCalls.ptrcallWithObjectAndIntArgRetVector2List(collideShapeBind, handle, parameters.requireOpenHandle(), maxResults)
     }
 
     /**
@@ -98,8 +98,8 @@ open class PhysicsDirectSpaceState2D(handle: MemorySegment) : GodotObject(handle
      *
      * Generated from Godot docs: PhysicsDirectSpaceState2D.get_rest_info
      */
-    fun getRestInfo(parameters: PhysicsShapeQueryParameters2D?): Map<String, Any?> {
-        return ObjectCalls.ptrcallWithObjectArgRetDictionary(getRestInfoBind, handle, parameters?.requireOpenHandle() ?: MemorySegment.NULL)
+    fun getRestInfo(parameters: PhysicsShapeQueryParameters2D): Map<String, Any?> {
+        return ObjectCalls.ptrcallWithObjectArgRetDictionary(getRestInfoBind, handle, parameters.requireOpenHandle())
     }
 
     companion object {